requirements.txt
torch
transformers>=4.38.0
gradio
Pillow
